Added a SecurityGroup for RemoteNode and RemotePod
If a VpcId is given, without a SecurityGroupId, a SecurityGroup will be
created. If a SecurityGroupId is given it will override the creation,
and that SecurityGroup will be used exclusively.

An EKS Cluster supporting Hybrid Nodes requires a security group that
allows port 443 from those cidrs. This commit adds a secondary security
group for those cidrs.

make security group creation optional

change condition to HasSecurityGroupId

Override security group cration if an id is given