Add command to ssh into e2e peered nodes
This just leverages the mechanism used by the e2e tests themselves to
get a shell into a e2e node. In order to not expose SSH to the internet
in those nodes and since SSM is not reliable, we leverage SSM in a
separate machine and use to jump to the target machine.